change method signature

+++ include/mgos_hlw8012.h
@@ -21,7 +21,7 @@ double mgos_hlw8012_readCurrent(HLW8012 *sensor);
 
 unsigned int mgos_hlw8012_readVoltage(HLW8012 *sensor);
 
-unsigned int mgos_hlw8012_readEnergy(HLW8012 *sensor);
+unsigned long mgos_hlw8012_readEnergy(HLW8012 *sensor);
 
 unsigned int mgos_hlw8012_readActivePower(HLW8012 *sensor);
